+++ samples/android-app/src/main/java/org/jlleitschuh/gradle/ktlint/android/ItemDetailFragment.kt
@@ -1,0 +1,58 @@
+package org.jlleitschuh.gradle.ktlint.android
+
+import android.os.Bundle
+import android.support.v4.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import kotlinx.android.synthetic.main.activity_item_detail.toolbar_layout
+import kotlinx.android.synthetic.main.item_detail.view.item_detail
+import org.jlleitschuh.gradle.ktlint.android.dummy.DummyContent
+
+/**
+ * A fragment representing a single Item detail screen.
+ * This fragment is either contained in a [ItemListActivity]
+ * in two-pane mode (on tablets) or a [ItemDetailActivity]
+ * on handsets.
+ */
+class ItemDetailFragment : Fragment() {
+
+    /**
+     * The dummy content this fragment is presenting.
+     */
+    private var mItem: DummyContent.DummyItem? = null
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        if (arguments?.containsKey(ARG_ITEM_ID) == true) {
+            // Load the dummy content specified by the fragment
+            // arguments. In a real-world scenario, use a Loader
+            // to load content from a content provider.
+            mItem = DummyContent.ITEM_MAP[arguments!!.getString(ARG_ITEM_ID)]
+            mItem?.let {
+                activity?.toolbar_layout?.title = it.content
+            }
+        }
+    }
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+        val rootView = inflater.inflate(R.layout.item_detail, container, false)
+
+        // Show the dummy content as text in a TextView.
+        mItem?.let {
+            rootView.item_detail.text = it.details
+        }
+
+        return rootView
+    }
+
+    companion object {
+        /**
+         * The fragment argument representing the item ID that this fragment
+         * represents.
+         */
+        const val ARG_ITEM_ID = "item_id"
+    }
+}