+++ samples/kotlin-gradle/src/main/kotlin/org/jlleitschuh/gradle/ktlint/sample/gradle/main.kt
@@ -1,0 +1,5 @@
+package org.jlleitschuh.gradle.ktlint.sample.gradle
+
+fun main(args: Array<String>) {
+    (1..10).map { "Value: $it" }.forEach { println("Gradle $it") }
+}